Fixed the alignment bug for w=64.  This will continue to be a problem with w=128, btw...
git-svn-id: svn://mamba.eecs.utk.edu/home/plank/svn/Galois-Library@82 36f187d4-5712-4624-889c-152d48957efa

+++ gf.c
@@ -330,7 +330,7 @@ void gf_set_region_data(gf_region_data *rd,
   void *src,
   void *dest,
   int bytes,
-  uint32_t val,
+  uint64_t val,
   int xor,
   int align)
 {
